fix[api]: rework handling of API request in aws_lambdas/submit_passport/submit_passport
